+++ python/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New public APIs: `identify_paths`, `identify_path`, `identify_bytes`.
 - Updated description for "unknown" to "Unknown binary data"
 - Magika CLI now defaults to "high-confidence" mode. "default" mode is now called "medium-confidence".
+- Magika CLI `--output-probability` has been renamed to `--output-score`.
 - Default model is now called `standard_v1`.
 - Major refactoring and clean up.